fix: moved function to another file
