+++ frontend/app/src/helpers/helpers-extended.ts
@@ -278,3 +278,24 @@ export const spliceOutPubkey = (userAddress: string): string => {
 
   return userAddress;
 };
+
+export function handleDisplayRole(displayedRoles: RolesCategory[]) {
+  // Set default data roles for first assign user
+  const defaultRole = {
+    'Manage bounties': true,
+    'Fund organization': true,
+    'Withdraw from organization': true,
+    'View transaction history': true
+  };
+
+  const tempDataRole: { [id: string]: boolean } = {};
+  const newDisplayedRoles = displayedRoles.map((role: RolesCategory) => {
+    if (defaultRole[role.name]) {
+      role.status = true;
+      role.roles.forEach((dataRole: string) => (tempDataRole[dataRole] = true));
+    }
+    return role;
+  });
+
+  return { newDisplayedRoles, tempDataRole };
+}
